backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
python-docx
python-multipart
boto3